Removes `crate::element::owned` module
